+++ index.d.ts
@@ -99,10 +99,10 @@ declare module 'mongoose' {
    */
   export function isValidObjectId(v: any): boolean;
 
-  export function model<T extends Document>(name: string, schema?: Schema, collection?: string, skipInit?: boolean): Model<T>;
+  export function model<T extends Document>(name: string, schema?: Schema<any>, collection?: string, skipInit?: boolean): Model<T>;
   export function model<T extends Document, U extends Model<T>>(
     name: string,
-    schema?: Schema,
+    schema?: Schema<T, U>,
     collection?: string,
     skipInit?: boolean
   ): U;
@@ -243,10 +243,10 @@ declare module 'mongoose' {
     models: { [index: string]: Model<any> };
 
     /** Defines or retrieves a model. */
-    model<T extends Document>(name: string, schema?: Schema, collection?: string): Model<T>;
+    model<T extends Document>(name: string, schema?: Schema<T>, collection?: string): Model<T>;
     model<T extends Document, U extends Model<T>>(
       name: string,
-      schema?: Schema,
+      schema?: Schema<T, U>,
       collection?: string,
       skipInit?: boolean
     ): U;
@@ -827,7 +827,10 @@ declare module 'mongoose' {
     /** Schema the model uses. */
     schema: Schema;
 
-    /** Creates a `update` query: updates one or many documents that match `filter` with `update`, based on the `multi` option. */
+    /**
+     * @deprecated use `updateOne` or `updateMany` instead.
+     * Creates a `update` query: updates one or many documents that match `filter` with `update`, based on the `multi` option.
+     */
     update(filter?: FilterQuery<T>, update?: UpdateQuery<T>, options?: QueryOptions | null, callback?: (err: any, res: any) => void): Query<any, T>;
 
     /** Creates a `updateMany` query: updates all documents that match `filter` with `update`. */
@@ -1030,7 +1033,7 @@ declare module 'mongoose' {
     useProjection?: boolean;
   }
 
-  class Schema<T = any> extends events.EventEmitter {
+  class Schema<DocType extends Document = Document, M extends Model<DocType> = Model<DocType>> extends events.EventEmitter {
     /**
      * Create a new schema
      */
@@ -1080,10 +1083,10 @@ declare module 'mongoose' {
     /** Adds an instance method to documents constructed from Models compiled from this schema. */
     // eslint-disable-next-line @typescript-eslint/ban-types
     method(name: string, fn: Function, opts?: any): this;
-    method(methods: any): this;
+    method(obj: { [name: string]: Function }): this;
 
     /** Object of currently defined methods on this schema. */
-    methods: any;
+    methods: { [name: string]: Function };
 
     /** The original object passed to the schema constructor */
     obj: any;
@@ -1104,21 +1107,21 @@ declare module 'mongoose' {
     plugin(fn: (schema: Schema, opts?: any) => void, opts?: any): this;
 
     /** Defines a post hook for the model. */
-    post<T extends Document = Document>(method: 'validate' | 'save' | 'remove' | 'updateOne' | 'deleteOne' | 'init' | RegExp, fn: (this: T, res: any, next: (err?: CallbackError) => void) => void): this;
+    post<T extends Document = DocType>(method: 'validate' | 'save' | 'remove' | 'updateOne' | 'deleteOne' | 'init' | RegExp, fn: (this: T, res: any, next: (err?: CallbackError) => void) => void): this;
     post<T extends Query<any, any> = Query<any, any>>(method: string | RegExp, fn: (this: T, res: any, next: (err: CallbackError) => void) => void): this;
     post<T extends Aggregate<any> = Aggregate<any>>(method: 'aggregate' | RegExp, fn: (this: T, res: Array<any>, next: (err: CallbackError) => void) => void): this;
-    post<T extends Model<any> = Model<any>>(method: 'insertMany' | RegExp, fn: (this: T, res: any, next: (err: CallbackError) => void) => void): this;
+    post<T extends Model<DocType> = M>(method: 'insertMany' | RegExp, fn: (this: T, res: any, next: (err: CallbackError) => void) => void): this;
 
-    post<T extends Document = Document>(method: 'validate' | 'save' | 'remove' | 'updateOne' | 'deleteOne' | 'init' | RegExp, fn: (this: T, err: NativeError, res: any, next: (err?: CallbackError) => void) => void): this;
+    post<T extends Document = DocType>(method: 'validate' | 'save' | 'remove' | 'updateOne' | 'deleteOne' | 'init' | RegExp, fn: (this: T, err: NativeError, res: any, next: (err?: CallbackError) => void) => void): this;
     post<T extends Query<any, any> = Query<any, any>>(method: string | RegExp, fn: (this: T, err: NativeError, res: any, next: (err: CallbackError) => void) => void): this;
     post<T extends Aggregate<any> = Aggregate<any>>(method: 'aggregate' | RegExp, fn: (this: T, err: NativeError, res: Array<any>, next: (err: CallbackError) => void) => void): this;
-    post<T extends Model<any> = Model<any>>(method: 'insertMany' | RegExp, fn: (this: T, err: NativeError, res: any, next: (err: CallbackError) => void) => void): this;
+    post<T extends Model<DocType> = M>(method: 'insertMany' | RegExp, fn: (this: T, err: NativeError, res: any, next: (err: CallbackError) => void) => void): this;
 
     /** Defines a pre hook for the model. */
-    pre<T extends Document = Document>(method: 'validate' | 'save' | 'remove' | 'updateOne' | 'deleteOne' | 'init' | RegExp, fn: (this: T, next: (err?: CallbackError) => void) => void): this;
+    pre<T extends Document = DocType>(method: 'validate' | 'save' | 'remove' | 'updateOne' | 'deleteOne' | 'findOneAndUpdate' | 'init' | RegExp, fn: (this: T, next: (err?: CallbackError) => void) => void): this;
     pre<T extends Query<any, any> = Query<any, any>>(method: string | RegExp, fn: (this: T, next: (err: CallbackError) => void) => void): this;
     pre<T extends Aggregate<any> = Aggregate<any>>(method: 'aggregate' | RegExp, fn: (this: T, next: (err: CallbackError) => void) => void): this;
-    pre<T extends Model<any> = Model<any>>(method: 'insertMany' | RegExp, fn: (this: T, next: (err: CallbackError) => void) => void): this;
+    pre<T extends Model<DocType> = M>(method: 'insertMany' | RegExp, fn: (this: T, next: (err: CallbackError) => void) => void): this;
 
     /** Object of currently defined query helpers on this schema. */
     query: any;
@@ -1142,7 +1145,7 @@ declare module 'mongoose' {
     static(obj: { [name: string]: Function }): this;
 
     /** Object of currently defined statics on this schema. */
-    statics: any;
+    statics: { [name: string]: Function };
 
     /** Creates a virtual type with the given name. */
     virtual(name: string, options?: any): VirtualType;
@@ -1928,7 +1931,7 @@ declare module 'mongoose' {
     j(val: boolean | null): this;
 
     /** Sets the lean option. */
-    lean(val?: boolean | any): Query<LeanDocumentOrArray<ResultType>, DocType>;
+    lean<LeanResultType = LeanDocumentOrArray<ResultType>>(val?: boolean | any): Query<LeanResultType, DocType>;
 
     /** Specifies the maximum number of documents the query will return. */
     limit(val: number): this;